pyproject.toml
[project]
name = "simple-mcp-server"
version = "0.1.0"
description = "Simple MCP server for Databricks Apps"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.12",
  "mcp[cli]>=1.14.0",
  "uvicorn>=0.34.2",
  "pydantic>=2",
  "fastmcp>=2.12.5",
  "python-dotenv>=1.0.0",
  "scalekit-sdk-python>=2.3.2",
]

[project.scripts]
simple-mcp-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
